Moved the maintainer package one level up
