+++ cmd/cmd.go
@@ -35,7 +35,7 @@ func init() {
 		StartCommand,
 		PingCommand,
 		EthereumCommand,
-		TbtcMaintainerCommand,
+		MaintainerCommand,
 	)
 }
