+++ cmd/flags.go
@@ -45,6 +45,8 @@ func initFlags(
 			initClientInfoFlags(cmd, cfg)
 		case config.Tbtc:
 			initTbtcFlags(cmd, cfg)
+		case config.Maintainer:
+			initMaintainerFlags(cmd, cfg)
 		case config.Developer:
 			initDeveloperFlags(cmd)
 		}
@@ -271,9 +273,12 @@ func initTbtcFlags(cmd *cobra.Command, cfg *config.Config) {
 		tbtc.DefaultKeyGenerationConcurrency,
 		"tECDSA key generation concurrency.",
 	)
+}
 
-	cmd.Flags().BoolVar(
-		&cfg.Tbtc.Maintainer.BitcoinDifficulty,
+// Initialize flags for Maintainer configuration.
+func initMaintainerFlags(command *cobra.Command, cfg *config.Config) {
+	command.Flags().BoolVar(
+		&cfg.Maintainer.BitcoinDifficulty,
 		"bitcoinDifficulty",
 		false,
 		"start Bitcoin difficulty maintainer",
