+++ test/config.json
@@ -34,10 +34,8 @@
         "NetworkMetricsTick": "43s",
         "EthereumMetricsTick": "1m27s"
     },
-    "Tbtc": {
-        "Maintainer": {
-            "BitcoinDifficulty": true
-        }
+    "Maintainer": {
+        "BitcoinDifficulty": true
     },
     "Developer": {
         "RandomBeaconAddress": "0xcf64c2a367341170cb4e09cf8c0ed137d8473ceb",
