+++ test/config.yaml
@@ -25,8 +25,7 @@ ClientInfo:
   Port: 3498
   NetworkMetricsTick: "43s"
   EthereumMetricsTick: "1m27s"
-Tbtc:
-  Maintainer:
+Maintainer:
     BitcoinDifficulty: true
 Developer:
   RandomBeaconAddress: "0xcf64c2a367341170cb4e09cf8c0ed137d8473ceb"
